Correcting one sonarqube violation.

+++ org.sat4j.maxsat/src/main/java/org/sat4j/maxsat/MinCostDecorator.java
@@ -67,7 +67,7 @@ public class MinCostDecorator extends PBSolverDecorator implements
 
     private final IVec<BigInteger> coeffs = new Vec<BigInteger>();
 
-    private int objectivevalue;
+    private long objectivevalue;
 
     private IConstr prevConstr;
